feat(tag-recipes): rename middleware to middlewares
