+++ server/routes/recipe-routes.js
@@ -1,6 +1,6 @@
 import { Router } from 'express';
 import controller from '../controllers/recipe-controller';
-import UsersMiddleware from '../middlewares';
+import UsersMiddleware from '../middlewares/users-middleware';
 import { loadRecipe, isRecipeOwner } from '../middlewares/recipe-middleware';
 
 const router = Router();
